# enterprise-vision-mlops/contracts/control-panel/control-panel.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enterprise Vision MLOps Control Panel API",
    "version": "2026-07-12.post-w7.multi-cycle-release.v1",
    "description": "Multi-cycle metadata and guarded mutation contract for the enterprise MLOps Control Panel, including runtime task dispatch, CI evidence, deployment intent state transitions, organizational service scope, drift review, CD/CT gates, and real-test Torch EfficientNet model matrices."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "local FastAPI service"
    }
  ],
  "paths": {
...
    "/control-panel/v1/tasks": {
      "get": {
        "summary": "Return the local task assignment ledger.",
        "operationId": "listTaskAssignments",
        "responses": {
          "200": {
            "description": "task assignment list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["tasks"],
                  "properties": {
                    "tasks": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/TaskAssignment"}
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Draft or queue an Airflow/MLflow-aware task assignment.",
        "operationId": "createTaskAssignment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/TaskAssignmentRequest"}
            }
          }
        },
        "responses": {
          "202": {
            "description": "task assignment accepted",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/TaskAssignment"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"}
        }
      }
    },
...
    "/control-panel/v1/tasks/{task_id}/dispatch": {
      "post": {
        "summary": "Dispatch a queued task through its registered runtime adapter.",
        "operationId": "dispatchTaskAssignment",
        "parameters": [
          {"name": "task_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "202": {
            "description": "task dispatched to the runtime",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskAssignment"}}}
          },
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"description": "task cannot be dispatched by the selected runtime adapter"}
        }
      }
    },
    "/control-panel/v1/tasks/{task_id}/confirm": {
      "post": {
        "summary": "Confirm a manual task assignment before runtime dispatch.",
        "operationId": "confirmTaskAssignment",
        "parameters": [
          {"name": "task_id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskTransitionRequest"}}}
        },
        "responses": {
          "202": {
            "description": "manual task confirmation recorded",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskAssignment"}}}
          },
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"description": "task state or external approval policy rejected confirmation"}
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "invalid request",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        }
      },
      "NotFound": {
        "description": "resource not found",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ErrorResponse"}
          }
        }
      }
    },
    "schemas": {
...
      "EnvironmentRef": {
        "type": "object",
        "required": ["name", "tier", "promotion_state"],
        "properties": {
          "name": {"type": "string"},
          "tier": {"type": "string", "enum": ["dev", "test", "staging", "pre-production", "production"]},
          "promotion_state": {"type": "string", "enum": ["draft", "candidate", "approved", "blocked", "deployed", "rolled_back"]},
          "cluster": {"type": ["string", "null"]},
          "namespace": {"type": ["string", "null"]},
          "release_ref": {"type": ["string", "null"]},
          "approval_policy": {"type": ["string", "null"]},
          "promotion_blockers": {
            "type": "array",
            "items": {"type": "string"}
          }
        }
      },
...
      "CDCTGate": {
        "type": "object",
        "required": ["status", "ci_status", "cd_status", "ct_status", "required_checks"],
        "properties": {
          "status": {"$ref": "#/components/schemas/State"},
          "ci_status": {"$ref": "#/components/schemas/State"},
          "cd_status": {"$ref": "#/components/schemas/State"},
          "ct_status": {"$ref": "#/components/schemas/State"},
          "required_checks": {
            "type": "array",
            "items": {"type": "string"}
          },
          "passed_checks": {
            "type": "array",
            "items": {"type": "string"}
          },
          "failed_checks": {
            "type": "array",
            "items": {"type": "string"}
          },
          "pipeline_run_uri": {"type": ["string", "null"]},
          "ct_trigger": {"type": ["string", "null"], "enum": ["new-data", "new-code", "drift", "schedule", "manual", null]},
          "approved_by": {"type": ["string", "null"]},
          "promotion_blockers": {
            "type": "array",
            "items": {"type": "string"}
          },
          "gate_report_uri": {"type": ["string", "null"]},
          "promotion_decision": {"type": "string", "enum": ["allow", "block", "manual_review"]},
          "block_reason": {"type": ["string", "null"]},
          "verification_summary": {
            "type": "object",
            "additionalProperties": {"$ref": "#/components/schemas/State"}
          }
        }
      },
...
      "TaskAssignmentRequest": {
        "type": "object",
        "required": ["task_type", "owner", "priority", "resource_profile", "config_payload"],
        "properties": {
          "cycle_id": {"type": ["string", "null"]},
          "task_type": {"type": "string", "enum": ["airflow_dag_run", "mlflow_run", "kubernetes_job"]},
          "owner": {"type": "string"},
          "priority": {"type": "string", "enum": ["low", "normal", "high", "urgent"]},
          "resource_profile": {"type": "string"},
          "requester_team": {"type": ["string", "null"]},
          "environment": {"$ref": "#/components/schemas/EnvironmentRef"},
          "approval_policy": {"type": ["string", "null"]},
          "airflow": {"$ref": "#/components/schemas/AirflowRef"},
          "mlflow": {"$ref": "#/components/schemas/MLflowRef"},
          "cdct_gate": {"$ref": "#/components/schemas/CDCTGate"},
          "config_payload": {"type": "object"},
          "dry_run": {"type": "boolean", "default": true}
        }
      },
      "TaskAssignment": {
        "allOf": [
          {"$ref": "#/components/schemas/TaskAssignmentRequest"},
          {
            "type": "object",
            "required": ["task_id", "status", "created_at", "audit"],
            "properties": {
              "task_id": {"type": "string"},
              "status": {"type": "string", "enum": ["draft", "dry_run", "queued", "pending_confirmation", "running", "done", "failed", "cancelled", "blocked"]},
              "created_at": {"type": "string", "format": "date-time"},
              "queued_at": {"type": ["string", "null"], "format": "date-time"},
              "dispatched_at": {"type": ["string", "null"], "format": "date-time"},
              "finished_at": {"type": ["string", "null"], "format": "date-time"},
              "runtime_system": {"type": ["string", "null"]},
              "runtime_id": {"type": ["string", "null"]},
              "runtime_url": {"type": ["string", "null"]},
              "runtime_state": {"type": ["string", "null"]},
              "runtime_evidence_uri": {"type": ["string", "null"]},
              "failure_reason": {"type": ["string", "null"]},
              "audit": {
                "type": "array",
                "items": {"$ref": "#/components/schemas/AuditEvent"}
              }
            }
          }
        ]
      },
      "TaskTransitionRequest": {
        "type": "object",
        "required": ["actor", "reason"],
        "properties": {
          "actor": {"type": "string"},
          "reason": {"type": "string"}
        }
      },
...
      "AuditEvent": {
        "type": "object",
        "required": ["timestamp", "actor", "event"],
        "properties": {
          "timestamp": {"type": "string", "format": "date-time"},
          "actor": {"type": "string"},
          "event": {"type": "string"},
          "details": {"type": "object"}
        }
      },
      "AirflowRef": {
        "type": "object",
        "properties": {
          "mode": {"type": "string", "enum": ["external-compose", "in-cluster", "managed-service"]},
          "control_mode": {"type": "string", "enum": ["read-only", "rest-api", "cli-bridge", "kubernetes-operator"]},
          "dag_id": {"type": "string"},
          "dag_run_id": {"type": "string"},
          "task_id": {"type": ["string", "null"]},
          "namespace": {"type": ["string", "null"]},
          "deployment_name": {"type": ["string", "null"]},
          "contract_config_map": {"type": ["string", "null"]},
          "connection_status": {"$ref": "#/components/schemas/State"},
          "url": {"type": ["string", "null"]}
        }
      },
...
      "MLflowRef": {
        "type": "object",
        "properties": {
          "experiment_id": {"type": "string"},
          "run_id": {"type": "string"},
          "model_name": {"type": "string"},
          "model_version": {"type": "string"},
          "url": {"type": ["string", "null"]}
        }
      },
...
      "State": {
        "type": "string",
        "enum": ["unknown", "queued", "running", "pass", "warn", "fail", "blocked", "done", "cancelled"]
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": {"type": "string"},
          "message": {"type": "string"},
          "details": {"type": "object"}
        }
      }
    }
  }
}